The declaration of CrosLog structures has been moved to cros_loh.h since they are mainly used in cros_log.c
